@@ -4,13 +4,13 @@ using Microsoft.Azure.WebJobs.Hosting;
 
 [assembly: WebJobsStartup(typeof(SqsExtensionStartup))]
 
-namespace Azure.Functions.Extensions.SQS
+namespace Azure.Functions.Extensions.SQS;
+
+public class SqsExtensionStartup : IWebJobsStartup
 {
-    public class SqsExtensionStartup : IWebJobsStartup
+    public void Configure(IWebJobsBuilder builder)
     {
-        public void Configure(IWebJobsBuilder builder)
-        {
-            builder.AddExtension<SqsExtensionProvider>().BindOptions<SqsQueueOptions>();
-        }
+        builder.AddExtension<SqsExtensionProvider>()
+               .BindOptions<SqsQueueOptions>();
     }
 }
